chore: move some stuff from fedimint-ln-client to fedimint-ln-common
This will help make them usable by the new gateway client.
