+++ modules/fedimint-ln-client/src/lib.rs
@@ -1,4 +1,3 @@
-mod api;
 mod db;
 pub mod pay;
 pub mod receive;
@@ -8,10 +7,8 @@ use std::sync::Arc;
 use std::time::{Duration, SystemTime};
 
 use anyhow::{bail, format_err};
-use api::LnFederationApi;
 use async_stream::stream;
 use bitcoin::{KeyPair, Network};
-use bitcoin_hashes::sha256::Hash as Sha256Hash;
 use bitcoin_hashes::Hash;
 use db::LightningGatewayKey;
 use fedimint_client::derivable_secret::DerivableSecret;
@@ -33,9 +30,12 @@ use fedimint_core::module::{
     CommonModuleGen, ExtendsCommonModuleGen, ModuleCommon, TransactionItemAmount,
 };
 use fedimint_core::{apply, async_trait_maybe_send, Amount, OutPoint, TransactionId};
+use fedimint_ln_common::api::LnFederationApi;
 use fedimint_ln_common::config::LightningClientConfig;
 use fedimint_ln_common::contracts::incoming::IncomingContractOffer;
-use fedimint_ln_common::contracts::outgoing::OutgoingContract;
+use fedimint_ln_common::contracts::outgoing::{
+    OutgoingContract, OutgoingContractAccount, OutgoingContractData,
+};
 use fedimint_ln_common::contracts::{Contract, EncryptedPreimage, IdentifiableContract, Preimage};
 pub use fedimint_ln_common::*;
 use futures::StreamExt;
@@ -44,9 +44,7 @@ use lightning::ln::PaymentSecret;
 use lightning::routing::gossip::RoutingFees;
 use lightning::routing::router::{RouteHint, RouteHintHop};
 use lightning_invoice::{Currency, Invoice, InvoiceBuilder, DEFAULT_EXPIRY_TIME};
-use pay::{
-    GatewayPayError, LightningPayStateMachine, OutgoingContractAccount, OutgoingContractData,
-};
+use pay::{GatewayPayError, LightningPayStateMachine};
 use rand::{CryptoRng, Rng, RngCore};
 use receive::{LightningReceiveError, LightningReceiveStateMachine};
 use secp256k1_zkp::{All, Secp256k1};
