+++ modules/fedimint-ln-common/src/lib.rs
@@ -9,6 +9,7 @@
 
 extern crate core;
 
+pub mod api;
 pub mod config;
 pub mod contracts;
 pub mod db;
